fixed standings (mostly), fixed rosters (mostly) see below:
Fixed some data fetching for standings and rosters, though a more full fix needs to be done once i run the data generation again and put the ea id in:
The reason for this is because no ea id on the matches i causing them to not be calculated and thus standings can't have it's stats and players can't have theirs.

Added more styling:
Added logos to standings and roster pages. Fixed the rosters scroll-to functionality to have better offset. Added a shimmer effect to the team scrolled to so it's clear which one to look at.

Note: CHL has not been working yet need to investigate why.
AHL, ECHL, CHL divisions/ conferences not being fetched, need to investigate that.

Some logos on the AHL side (Henderson, Calagry, Cleveland, etc.) are the correct size but others are smaller. Need to look into a fix or replacing some logos.